Convert to ts files and export as default
