+++ playbook/app/javascript/theme.js
@@ -3,5 +3,5 @@ export {default as mapTheme} from '../pb_kits/playbook/pb_map/pbMapTheme'
 export {default as useCollapsible} from '../pb_kits/playbook/pb_collapsible/useCollapsible'
 export {default as useDropdown } from '../pb_kits/playbook/pb_dropdown/hooks/useDropdown'
 
-export { highchartsTheme } from '../pb_kits/playbook/pb_dashboard/pbChartsLightTheme'
-export { highchartsDarkTheme } from '../pb_kits/playbook/pb_dashboard/pbChartsDarkTheme'
+export {default as highchartsTheme} from '../pb_kits/playbook/pb_dashboard/pbChartsLightTheme'
+export {default as highchartsDarkTheme } from '../pb_kits/playbook/pb_dashboard/pbChartsDarkTheme'
