+++ playbook/app/pb_kits/playbook/pb_dashboard/pbChartsDarkTheme.ts
@@ -1,7 +1,7 @@
 import colors from '../tokens/exports/_colors.module.scss'
 import typography from '../tokens/exports/_typography.module.scss'
 
-const highchartsTheme = {
+const highchartsDarkTheme = {
   lang: {
     thousandsSep: ',',
   },
@@ -17,13 +17,12 @@ const highchartsTheme = {
   chart: {
     borderWidth: 0,
     borderRadius: 0,
-    plotBackgroundColor: undefined,
     plotShadow: false,
     plotBorderWidth: 0,
   },
   title: {
     style: {
-      color: colors.text_lt_default,
+      color: colors.text_dk_default,
       fontFamily: typography.font_family_base,
       fontWeight: typography.bold,
       fontSize: typography.heading_3,
@@ -32,26 +31,26 @@ const highchartsTheme = {
   subtitle: {
     style: {
       fontFamily: typography.font_family_base,
-      color: colors.text_lt_light,
+      color: colors.text_dk_light,
       fontWeight: typography.regular,
       fontSize: typography.text_base,
     },
   },
   xAxis: {
     gridLineWidth: 0,
-    lineColor: colors.border_light,
-    tickColor: colors.border_light,
+    lineColor: colors.border_dark,
+    tickColor: colors.border_dark,
     labels: {
       style: {
         fontFamily: typography.font_family_base,
-        color: colors.text_lt_lighter,
+        color: colors.text_dk_lighter,
         fontWeight: typography.bold,
         fontSize: typography.text_smaller,
       },
     },
     title: {
       style: {
-        color: colors.text_lt_default,
+        color: colors.text_dk_default,
         fontFamily: typography.font_family_base,
         fontWeight: typography.regular,
         fontSize: typography.heading_4,
@@ -59,16 +58,14 @@ const highchartsTheme = {
     },
   },
   yAxis: {
-    alternateGridColor: undefined,
-    minorTickInterval: null,
-    gridLineColor: colors.border_light,
-    minorGridLineColor: colors.border_light,
+    gridLineColor: colors.border_dark,
+    minorGridLineColor: colors.border_dark,
     lineWidth: 0,
     tickWidth: 0,
     labels: {
       style: {
         fontFamily: typography.font_family_base,
-        color: colors.text_lt_lighter,
+        color: colors.text_dk_lighter,
         fontWeight: typography.bold,
         fontSize: typography.text_smaller,
       },
@@ -76,7 +73,7 @@ const highchartsTheme = {
     title: {
       style: {
         fontFamily: typography.font_family_base,
-        color: colors.text_lt_lighter,
+        color: colors.text_dk_lighter,
         fontWeight: typography.bold,
         fontSize: typography.text_smaller,
       },
@@ -88,23 +85,23 @@ const highchartsTheme = {
     verticalAlign: 'bottom',
     itemStyle: {
       fontFamily: typography.font_family_base,
-      color: colors.text_lt_light,
+      color: colors.text_dk_light,
       fontWeight: typography.regular,
       fontSize: typography.text_smaller,
     },
     itemHoverStyle: {
-      color: colors.text_lt_default,
+      color: colors.text_dk_default,
     },
     itemHiddenStyle: {
-      color: colors.text_lt_lighter,
+      color: colors.text_dk_lighter,
     },
   },
   tooltip: {
     backgroundColor: {
       linearGradient: { x1: 0, y1: 0, x2: 0, y2: 1 },
       stops: [
-        [0, colors.bg_dark],
-        [1, colors.bg_dark],
+        [0, colors.card_light],
+        [1, colors.card_light],
       ],
     },
     shadow: false,
@@ -112,7 +109,7 @@ const highchartsTheme = {
     borderRadius: 10,
     style: {
       fontFamily: typography.font_family_base,
-      color: colors.text_dk_default,
+      color: colors.text_lt_default,
       fontWeight: typography.regular,
       fontSize: typography.text_smaller,
     },
@@ -123,13 +120,14 @@ const highchartsTheme = {
     background: {
       // eslint-disable-next-line @typescript-eslint/ban-ts-comment 
       // @ts-ignore
-      borderColor: colors.border_light, 
+      borderColor: colors.border_dark,
     },
   },
 
   plotOptions: {
     series: {
-      threshold: null,
+      borderColor: colors.bg_dark_card,
+      borderWidth: 2,
     },
     // PIE STYLES
     pie: {
@@ -146,9 +144,8 @@ const highchartsTheme = {
         style: {
           fontFamily: typography.font_family_base,
           fontSize: typography.text_smaller,
-          color: colors.text_lt_light,
+          color: colors.text_dk_light,
           fontWeight: typography.regular,
-          textOutline: '2px $white',
         },
       },
     },
@@ -156,10 +153,10 @@ const highchartsTheme = {
     // LINE CHART STYLES
     line: {
       dataLabels: {
-        color: '#CCC',
+        color: colors.text_dk_light,
       },
       marker: {
-        lineColor: '#333',
+        lineColor: colors.border_dark,
       },
     },
 
@@ -204,4 +201,4 @@ const highchartsTheme = {
   },
 }
 
-export { highchartsTheme }
+export default highchartsDarkTheme
