+++ playbook/app/pb_kits/playbook/pb_dashboard/pbChartsLightTheme.ts
@@ -1,7 +1,7 @@
 import colors from '../tokens/exports/_colors.module.scss'
 import typography from '../tokens/exports/_typography.module.scss'
 
-const highchartsDarkTheme = {
+const highchartsTheme = {
   lang: {
     thousandsSep: ',',
   },
@@ -17,13 +17,12 @@ const highchartsDarkTheme = {
   chart: {
     borderWidth: 0,
     borderRadius: 0,
-    plotBackgroundColor: undefined,
     plotShadow: false,
     plotBorderWidth: 0,
   },
   title: {
     style: {
-      color: colors.text_dk_default,
+      color: colors.text_lt_default,
       fontFamily: typography.font_family_base,
       fontWeight: typography.bold,
       fontSize: typography.heading_3,
@@ -32,26 +31,26 @@ const highchartsDarkTheme = {
   subtitle: {
     style: {
       fontFamily: typography.font_family_base,
-      color: colors.text_dk_light,
+      color: colors.text_lt_light,
       fontWeight: typography.regular,
       fontSize: typography.text_base,
     },
   },
   xAxis: {
     gridLineWidth: 0,
-    lineColor: colors.border_dark,
-    tickColor: colors.border_dark,
+    lineColor: colors.border_light,
+    tickColor: colors.border_light,
     labels: {
       style: {
         fontFamily: typography.font_family_base,
-        color: colors.text_dk_lighter,
+        color: colors.text_lt_lighter,
         fontWeight: typography.bold,
         fontSize: typography.text_smaller,
       },
     },
     title: {
       style: {
-        color: colors.text_dk_default,
+        color: colors.text_lt_default,
         fontFamily: typography.font_family_base,
         fontWeight: typography.regular,
         fontSize: typography.heading_4,
@@ -59,16 +58,14 @@ const highchartsDarkTheme = {
     },
   },
   yAxis: {
-    alternateGridColor: undefined,
-    minorTickInterval: null,
-    gridLineColor: colors.border_dark,
-    minorGridLineColor: colors.border_dark,
+    gridLineColor: colors.border_light,
+    minorGridLineColor: colors.border_light,
     lineWidth: 0,
     tickWidth: 0,
     labels: {
       style: {
         fontFamily: typography.font_family_base,
-        color: colors.text_dk_lighter,
+        color: colors.text_lt_lighter,
         fontWeight: typography.bold,
         fontSize: typography.text_smaller,
       },
@@ -76,7 +73,7 @@ const highchartsDarkTheme = {
     title: {
       style: {
         fontFamily: typography.font_family_base,
-        color: colors.text_dk_lighter,
+        color: colors.text_lt_lighter,
         fontWeight: typography.bold,
         fontSize: typography.text_smaller,
       },
@@ -88,23 +85,23 @@ const highchartsDarkTheme = {
     verticalAlign: 'bottom',
     itemStyle: {
       fontFamily: typography.font_family_base,
-      color: colors.text_dk_light,
+      color: colors.text_lt_light,
       fontWeight: typography.regular,
       fontSize: typography.text_smaller,
     },
     itemHoverStyle: {
-      color: colors.text_dk_default,
+      color: colors.text_lt_default,
     },
     itemHiddenStyle: {
-      color: colors.text_dk_lighter,
+      color: colors.text_lt_lighter,
     },
   },
   tooltip: {
     backgroundColor: {
       linearGradient: { x1: 0, y1: 0, x2: 0, y2: 1 },
       stops: [
-        [0, colors.card_light],
-        [1, colors.card_light],
+        [0, colors.bg_dark],
+        [1, colors.bg_dark],
       ],
     },
     shadow: false,
@@ -112,7 +109,7 @@ const highchartsDarkTheme = {
     borderRadius: 10,
     style: {
       fontFamily: typography.font_family_base,
-      color: colors.text_lt_default,
+      color: colors.text_dk_default,
       fontWeight: typography.regular,
       fontSize: typography.text_smaller,
     },
@@ -123,16 +120,11 @@ const highchartsDarkTheme = {
     background: {
       // eslint-disable-next-line @typescript-eslint/ban-ts-comment 
       // @ts-ignore
-      borderColor: colors.border_dark,
+      borderColor: colors.border_light, 
     },
   },
 
   plotOptions: {
-    series: {
-      borderColor: colors.bg_dark_card,
-      borderWidth: 2,
-      threshold: null,
-    },
     // PIE STYLES
     pie: {
       colors: [
@@ -148,8 +140,9 @@ const highchartsDarkTheme = {
         style: {
           fontFamily: typography.font_family_base,
           fontSize: typography.text_smaller,
-          color: colors.text_dk_light,
+          color: colors.text_lt_light,
           fontWeight: typography.regular,
+          textOutline: '2px $white',
         },
       },
     },
@@ -157,10 +150,10 @@ const highchartsDarkTheme = {
     // LINE CHART STYLES
     line: {
       dataLabels: {
-        color: colors.text_dk_light,
+        color: '#CCC',
       },
       marker: {
-        lineColor: colors.border_dark,
+        lineColor: '#333',
       },
     },
 
@@ -205,4 +198,4 @@ const highchartsDarkTheme = {
   },
 }
 
-export { highchartsDarkTheme }
+export default highchartsTheme
